website/backend/requirements.txt
fastapi 
uvicorn 
httpx 
pytest
SQLAlchemy
pydantic
python-multipart
PyPDF2
flair
torch
spacy
transformers
pandas
numpy
matplotlib

